+++ CHANGELOG.md
@@ -24,6 +24,7 @@ REMOVED:
 FIXED:
 
 - `WHERE` conditions won't be erased when using `SelectQueryBuilder::fetchById()`
+- Fetching relationships of an empty list of entities won't raise syntax error
 
 ## 0.2.0 - 2017-02-17